refacto(LEE): rename Entreprise en EntrepriseSouhaitantSEngager et suppression du doublon
