+++ src/client/services/lesEntreprisesSEngagent/lesEntreprisesSEngagentService.fixture.ts
@@ -1,5 +1,4 @@
-import { EntrepriseSouhaitantSEngager } from '~/pages/les-entreprises-s-engagent/inscription/index.page';
-import { Entreprise } from '~/server/entreprises/domain/Entreprise';
+import { EntrepriseSouhaitantSEngager } from '~/server/entreprises/domain/EntrepriseSouhaitantSEngager';
 import { createSuccess } from '~/server/errors/either';
 
 import { LesEntreprisesSEngagentService } from './lesEntreprisesSEngagent.service';
@@ -11,23 +10,6 @@ export function aLesEntreprisesSEngagentService(): LesEntreprisesSEngagentServic
 }
 
 export function anEntrepriseSouhaitantSEngager(overrides?: Partial<EntrepriseSouhaitantSEngager>): EntrepriseSouhaitantSEngager {
-	return {
-		codePostal: '75015',
-		email: 'toto@email.com',
-		nom: 'Tata',
-		nomSociété: 'Octo',
-		prénom: 'Toto',
-		secteur: 'health-social',
-		siret: '41816609600069',
-		taille: 'xsmall',
-		travail: 'RH',
-		téléphone: '0122334455',
-		ville: 'Paris 15e Arrondissement',
-		...overrides,
-	};
-}
-
-export function anEntreprise(overrides?: Partial<Entreprise>): Entreprise {
 	return {
 		codePostal: '75015',
 		email: 'machin.chose@bidule.com',
